refactor (sidebar>layout.tsx): - moved the client side interactivity down the component tree - refatored LeftSideBar componet for cleaner code - made a separate CreateFormButton and LeftSideBarNavLinks components
