+++ src/app/(with-sidebar)/layout.tsx
@@ -1,20 +1,10 @@
-"use client";
-import React, { useState, type ReactNode } from "react";
+import React, { type ReactNode } from "react";
 import { Grid2 as Grid, Box } from "@mui/material";
 import RightSideBar from "~/app/_components/sidebar/RightSideBar";
 import { MENU_ITEMS } from "~/app/(with-sidebar)/myfeed/static/menu";
 import LeftSideBar from "../_components/sidebar/LeftSideBar";
-import CreatePostForm from "../editor/_components/createPostForm/CreatePostForm";
 
 const Layout = ({ children }: { children: ReactNode }) => {
-  const [createPostFormOpen, setCreatePostFormOpen] = useState(false);
-  const handleCreatePostFormClose = () => {
-    setCreatePostFormOpen(false);
-  };
-  const handleCreatePostFormOpen = () => {
-    setCreatePostFormOpen(true);
-  };
-
   return (
     <Box
       sx={{
@@ -44,10 +34,7 @@ const Layout = ({ children }: { children: ReactNode }) => {
             mr: 4,
           }}
         >
-          <LeftSideBar
-            menuItems={MENU_ITEMS}
-            createPostFormOpen={handleCreatePostFormOpen}
-          />
+          <LeftSideBar menuItems={MENU_ITEMS} />
         </Grid>
         <Grid
           size={7}
@@ -72,20 +59,6 @@ const Layout = ({ children }: { children: ReactNode }) => {
           <RightSideBar />
         </Grid>
       </Grid>
-      {createPostFormOpen && (
-        <CreatePostForm
-          open={createPostFormOpen}
-          handleClose={handleCreatePostFormClose}
-          createPostFormData={{
-            title: "",
-            targetAudience: [],
-            thumbnailUrl: "",
-            postType: "",
-            tags: [],
-            actors: [],
-          }}
-        />
-      )}
     </Box>
   );
 };
