+++ src/app/(with-sidebar)/loading.tsx
@@ -1,0 +1,7 @@
+import Loader from "../_components/loader/Loader";
+
+const SideBarGroupLoader = () => {
+  return <Loader title="Loading..." height="100%" width="100%" />;
+};
+
+export default SideBarGroupLoader;
